Fixed some bugs after change to only abbreviations. Ongoing work

+++ src/main/java/se/sics/ace/as/Token.java
@@ -276,7 +276,7 @@ public class Token implements Endpoint, AutoCloseable {
 		
 		//6. Create token
 		//Find supported token type
-		Integer tokenType = null;
+		Short tokenType = null;
         try {
             tokenType = this.db.getSupportedTokenType(aud);
         } catch (AceException e) {
